# global-payments-platform/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Global Payments Platform API
  version: 0.1.0
servers:
  - url: https://localhost:8443
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /auth/me:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
  /accounts/:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List accounts
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOut'
  /transfers/:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List transfers
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCreate'
      responses:
        '200':
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferOut'
  /transfers/{id}/events:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferEventOut'
  /transfers/validate:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          text/xml:
            schema:
              type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateOut'
  /kyc/upload:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document_type:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded
  /notifications/stream:
    get:
      responses:
        '200':
          description: SSE stream
  /admin/status:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Status
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AccountCreate:
      type: object
      required: [iban, bic, display_name]
      properties:
        iban: { type: string }
        bic: { type: string }
        display_name: { type: string }
    AccountOut:
      type: object
      properties:
        id: { type: integer }
        iban: { type: string }
        bic: { type: string }
        display_name: { type: string }
    TransferCreate:
      type: object
      required: [debtor_account_id, creditor_iban, creditor_bic, amount, currency]
      properties:
        debtor_account_id: { type: integer }
        creditor_iban: { type: string }
        creditor_bic: { type: string }
        amount: { type: number, format: decimal }
        currency: { type: string }
        reference: { type: string }
        dry_run: { type: boolean, default: true }
    TransferOut:
      type: object
      properties:
        id: { type: integer }
        status: { type: string, enum: [INITIATED, PENDING, COMPLETED, FAILED] }
        reference: { type: string }
    TransferEventOut:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        payload: { type: string }
    ValidateOut:
      type: object
      properties:
        valid: { type: boolean }
        details: { type: string }